# public/docs/UberBoseOpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Überböse Streaming API",
    "description": "The missing API specifications for your streaming device.\n\nThis api spec combines the APIs of 4 different APIs / domains your speaker contacts.\n\n- **marge**: Account bound API managing source providers, presets, recents, and devices\n- **stats**: Device analytics and telemetry API collecting playback events and diagnostics\n- **software-update**: Firmware update information\n- **bmx**: API for streaming radio services (TuneIn, Custom Stations, etc.)\n",
    "version": "1.2.0"
  },
  "paths": {
    "/streaming/sourceproviders": {
      "get": {
        "summary": "Get streaming source providers",
        "description": "Retrieves a list of available streaming source providers",
        "operationId": "getSourceProviders",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with list of source providers",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/vnd.bose.streaming-v1.2+xml"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "1762017390476"
                }
              }
            },
            "content": {
              "application/vnd.bose.streaming-v1.2+xml": {
                "schema": {
                  "$ref": "#/components/schemas/SourceProvidersResponse"
                },
                "example": "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>\n<sourceProviders>\n  <sourceprovider id=\"1\">\n    <createdOn>2012-09-19T12:43:00.000+00:00</createdOn>\n    <name>PANDORA</name>\n    <updatedOn>2012-09-19T12:43:00.000+00:00</updatedOn>\n  </sourceprovider>\n  <sourceprovider id=\"15\">\n    <createdOn>2014-03-17T15:30:27.000+00:00</createdOn>\n    <name>SPOTIFY</name>\n    <updatedOn>2014-03-17T15:30:27.000+00:00</updatedOn>\n  </sourceprovider>\n</sourceProviders>\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing Bearer token"
          },
          "403": {
            "description": "Forbidden - Valid token but insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/streaming/account/{accountId}/device/{deviceId}/presets": {
      "get": {
        "summary": "Get device presets",
        "description": "Retrieves a list of presets configured for the device",
        "operationId": "getPresets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/accountId"
          },
          {
            "$ref": "#/components/parameters/deviceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with list of presets",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/vnd.bose.streaming-v1.2+xml"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "1736199076279"
                }
              }
            },
            "content": {
              "application/vnd.bose.streaming-v1.2+xml": {
                "schema": {
                  "$ref": "#/components/schemas/PresetsContainer"
                },
                "example": "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>\n<presets>\n  <preset buttonNumber=\"1\">\n    <containerArt>http://cdn-radiotime-logos.tunein.com/s80044q.png</containerArt>\n    <contentItemType>stationurl</contentItemType>\n    <createdOn>2018-11-26T18:40:45.000+00:00</createdOn>\n    <location>/v1/playback/station/s80044</location>\n    <name>Radio TEDDY</name>\n    <source id=\"19989313\" type=\"Audio\">\n      <createdOn>2018-08-11T08:55:41.000+00:00</createdOn>\n      <credential type=\"token\">eyJzZXJpYWwiOiAiZjUwY2JjZGItN2IxMi00NjJmLWE5ZTctZGEzMTg5NDcwMTU4In0=</credential>\n      <name></name>\n      <sourceproviderid>25</sourceproviderid>\n      <sourcename></sourcename>\n      <sourceSettings/>\n      <updatedOn>2019-07-20T17:48:31.000+00:00</updatedOn>\n      <username></username>\n    </source>\n    <updatedOn>2018-11-26T18:40:45.000+00:00</updatedOn>\n    <username>Radio TEDDY</username>\n  </preset>\n  <preset buttonNumber=\"2\">\n    <containerArt>https://mosaic.scdn.co/300/ab67616d0000b2732a4bc05d6d23b6a9a309e8f8ab67616d0000b2733d3c35cdbf9c51eefeaed7a7ab67616d0000b2738271a54c70aaae53e3da0c09ab67616d0000b273c4ec45569fc2db1edc44a465</containerArt>\n    <contentItemType>tracklisturl</contentItemType>\n    <createdOn>2018-11-14T18:27:39.000+00:00</createdOn>\n    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDoyM1NNZHlPSEE2S2t6SG9QT0o1S1E5</location>\n    <name>Radio Mix</name>\n    <source id=\"19989621\" type=\"Audio\">\n      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>\n      <credential type=\"token_version_3\">AQCfoupV5pdDJvqVjVdOF5tUOvyB3mKqeUGyZgqSYk4imOvta8p4dJnCHCNZANz-x0O-W2aQ15ML38pyx1CdvNSdOjGPwg0IlYYfPKWgoS07kmdEfRpf1Vur-OIx46KF56pjsA</credential>\n      <name>mock5zt8py3wuxy123xa431ge</name>\n      <sourceproviderid>15</sourceproviderid>\n      <sourcename>user@example.com</sourcename>\n      <sourceSettings/>\n      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>\n      <username>mock5zt8py3wuxy123xa431ge</username>\n    </source>\n    <updatedOn>2021-02-21T20:19:16.000+00:00</updatedOn>\n    <username>Radio Mix</username>\n  </preset>\n  <preset buttonNumber=\"3\">\n    <containerArt>http://cdn-profiles.tunein.com/s25111/images/logoq.jpg?t=1</containerArt>\n    <contentItemType>stationurl</contentItemType>\n    <createdOn>2018-11-26T18:41:18.000+00:00</createdOn>\n    <location>/v1/playback/station/s25111</location>\n    <name>radioeins vom rbb</name>\n    <source id=\"19989313\" type=\"Audio\">\n      <createdOn>2018-08-11T08:55:41.000+00:00</createdOn>\n      <credential type=\"token\">eyJzZXJpYWwiOiAiZjUwY2JjZGItN2IxMi00NjJmLWE5ZTctZGEzMTg5NDcwMTU4In0=</credential>\n      <name></name>\n      <sourceproviderid>25</sourceproviderid>\n      <sourcename></sourcename>\n      <sourceSettings/>\n      <updatedOn>2019-07-20T17:48:31.000+00:00</updatedOn>\n      <username></username>\n    </source>\n    <updatedOn>2018-11-26T18:41:18.000+00:00</updatedOn>\n    <username>radioeins vom rbb</username>\n  </preset>\n  <preset buttonNumber=\"4\">\n    <containerArt>https://mosaic.scdn.co/300/ab67616d0000b2735ca387be0b44d742bcee3317ab67616d0000b273627072ee659b0f3aaa0f537aab67616d0000b27364b153e7f4b060d9735c7d76ab67616d0000b2738ba1aaf88b157453fe12482d</containerArt>\n    <contentItemType>tracklisturl</contentItemType>\n    <createdOn>2018-11-26T18:47:06.000+00:00</createdOn>\n    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDoyd0JCOGIzUWhDWXd5T0d2dE9id3dI</location>\n    <name>Saisonal Simone Sommerland</name>\n    <source id=\"19989621\" type=\"Audio\">\n      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>\n      <credential type=\"token_version_3\">AQCfoupV5pdDJvqVjVdOF5tUOvyB3mKqeUGyZgqSYk4imOvta8p4dJnCHCNZANz-x0O-W2aQ15ML38pyx1CdvNSdOjGPwg0IlYYfPKWgoS07kmdEfRpf1Vur-OIx46KF56pjsA</credential>\n      <name>mock5zt8py3wuxy123xa431ge</name>\n      <sourceproviderid>15</sourceproviderid>\n      <sourcename>user@example.com</sourcename>\n      <sourceSettings/>\n      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>\n      <username>mock5zt8py3wuxy123xa431ge</username>\n    </source>\n    <updatedOn>2022-11-17T19:35:37.000+00:00</updatedOn>\n    <username>Saisonal Simone Sommerland</username>\n  </preset>\n  <preset buttonNumber=\"5\">\n    <containerArt></containerArt>\n    <contentItemType>tracklisturl</contentItemType>\n    <createdOn>2018-11-14T18:27:23.000+00:00</createdOn>\n    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDozN2k5ZFFaRjFFNG1zazBURmdzeVRG</location>\n    <name>Ed Sheeran Radio</name>\n    <source id=\"20260226\" type=\"Audio\">\n      <createdOn>2018-09-16T19:18:46.000+00:00</createdOn>\n      <credential type=\"token_version_3\">AQAHZ3A2USJzLX8CUEUPqIHOPpQihPpwcVO-0h2-g5-7SWRtIFfoQadeVu1Ud0TgVDkee7PyFnJbJUubZkrpv3A9jLaEeY991wSkXGG-EM_IlwvkRpASmwT3nbyixpVkSDw</credential>\n      <name>bldfvhhzdsyw2d4cy3melcaco</name>\n      <sourceproviderid>15</sourceproviderid>\n      <sourcename>otheruser@example.com</sourcename>\n      <sourceSettings/>\n      <updatedOn>2019-12-20T14:31:09.000+00:00</updatedOn>\n      <username>bldfvhhzdsyw2d4cy3melcaco</username>\n    </source>\n    <updatedOn>2020-01-26T12:45:13.000+00:00</updatedOn>\n    <username>Ed Sheeran Radio</username>\n  </preset>\n  <preset buttonNumber=\"6\">\n    <containerArt>https://image-cdn-fa.spotifycdn.com/image/ab67706c0000da843b38733ef58fbd3530776a42</containerArt>\n    <contentItemType>tracklisturl</contentItemType>\n    <createdOn>2018-08-11T09:53:48.000+00:00</createdOn>\n    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDoybjZXMnA1QzBNQUQ5YTR6NXhUVDdu</location>\n    <name>Komplett Entspannt</name>\n    <source id=\"19989621\" type=\"Audio\">\n      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>\n      <credential type=\"token_version_3\">AQCfoupV5pdDJvqVjVdOF5tUOvyB3mKqeUGyZgqSYk4imOvta8p4dJnCHCNZANz-x0O-W2aQ15ML38pyx1CdvNSdOjGPwg0IlYYfPKWgoS07kmdEfRpf1Vur-OIx46KF56pjsA</credential>\n      <name>mock5zt8py3wuxy123xa431ge</name>\n      <sourceproviderid>15</sourceproviderid>\n      <sourcename>user@example.com</sourcename>\n      <sourceSettings/>\n      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>\n      <username>mock5zt8py3wuxy123xa431ge</username>\n    </source>\n    <updatedOn>2025-01-06T21:31:15.000+00:00</updatedOn>\n    <username>Komplett Entspannt</username>\n  </preset>\n</presets>\n"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing Bearer token"
          },
          "403": {
            "description": "Forbidden - Valid token but insufficient permissions"
          },
          "404": {
            "description": "Not found - Account or device not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/streaming/account/{accountId}/full": {
      "get": {
        "summary": "Get full account details",
        "description": "Retrieves complete account information including devices, presets, recents, and sources",
        "operationId": "getFullAccount",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/accountId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with full account details",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/vnd.bose.streaming-v1.2+xml"
                }
              },
              "METHOD_NAME": {
                "schema": {
                  "type": "string",
                  "example": "getFullAccount"
                }
              }
            },
            "content": {
              "application/vnd.bose.streaming-v1.2+xml": {
                "schema": {
                  "$ref": "#/components/schemas/FullAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing Bearer token"
          },
          "403": {
            "description": "Forbidden - Valid token but insufficient permissions"
          },
          "404": {
            "description": "Not found - Account not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/streaming/account/{accountId}/provider_settings": {
      "get": {
        "summary": "Get provider settings",
        "description": "Unclear. Returns empty 200 ok response for me",
        "operationId": "getProviderSettings",
        "tags": [
          "experimental"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/accountId"
          }
        ],
        "responses": {
          "200": {
            "description": "Unclear, empty response",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/vnd.bose.streaming-v1.2+xml"
                }
              },
              "METHOD_NAME": {
                "schema": {
                  "type": "string",
                  "example": "getProviderSettings"
                }
              }
            }
          }
        }
      }
    },
...
    "/streaming/device/{deviceId}/streaming_token": {
      "get": {
        "summary": "Get streaming token",
        "description": "Unclear. Returns 200 ok response with empty body\nand an `Authorization` header with a value containing A-Z, a-z, 0-9, +, /\n",
        "operationId": "getStreamingToken",
        "tags": [
          "experimental"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/deviceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "headers": {
              "Authorization": {
                "schema": {
                  "type": "string",
                  "example": "1a2312e/423f243f+31f/qwe43r3"
                },
                "description": "Refreshed Bearer token"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing Bearer token"
          },
          "403": {
            "description": "Forbidden - Valid token but insufficient permissions"
          },
          "404": {
            "description": "Not found - Device not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token authentication"
      }
    },
    "parameters": {
      "accountId": {
        "name": "accountId",
        "in": "path",
        "required": true,
        "description": "Account identifier",
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-zA-Z-]+$",
          "example": "6921042"
        }
      },
      "deviceId": {
        "name": "deviceId",
        "in": "path",
        "required": true,
        "description": "Device identifier",
        "schema": {
          "type": "string",
          "example": "587A628A4042"
        }
      }
    },
    "schemas": {
      "SourceProvidersResponse": {
        "type": "object",
        "description": "XML response containing list of source providers",
        "xml": {
          "name": "sourceProviders"
        },
        "properties": {
          "sourceprovider": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceProvider"
            },
            "xml": {
              "wrapped": false
            }
          }
        }
      },
...
    }
  }
}